+++ pkg/register/register.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/ha1tch/repoman/pkg/config"
+	"github.com/ha1tch/gorepoman/pkg/config"
 )
 
 type Row struct {