+++ pkg/syncver/syncver.go
@@ -8,7 +8,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/ha1tch/repoman/pkg/config"
+	"github.com/ha1tch/gorepoman/pkg/config"
 )
 
 var versionRe = regexp.MustCompile(`^[0-9]+\.[0-9]+\.[0-9]+(-[a-zA-Z0-9.]+)?$`)